requirements.txt
psycopg2-binary
python-dotenv
lxml
beautifulsoup4 